Extract functionality for list-certificates into service class
